backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
pydantic-settings==2.1.0
sqlalchemy>=2.0.25
aiosqlite==0.19.0
